requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
paho-mqtt==2.1.0
passlib==1.7.4
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
